clang-tidy for rawfile.cpp

+++ core/include/aare/core/Frame.hpp
@@ -14,18 +14,18 @@ namespace aare {
  * should be able to work with streams coming from files or network
  */
 class Frame {
-    ssize_t m_rows;
-    ssize_t m_cols;
-    ssize_t m_bitdepth;
+    size_t m_rows;
+    size_t m_cols;
+    size_t m_bitdepth;
     std::byte *m_data;
 
   public:
-    Frame(ssize_t rows, ssize_t cols, ssize_t m_bitdepth);
-    Frame(std::byte *bytes, ssize_t rows, ssize_t cols, ssize_t m_bitdepth);
-    std::byte *get(int row, int col);
+    Frame(size_t rows, size_t cols, size_t m_bitdepth);
+    Frame(std::byte *bytes, size_t rows, size_t cols, size_t m_bitdepth);
+    std::byte *get(size_t row, size_t col);
 
     // TODO! can we, or even want to remove the template?
-    template <typename T> void set(int row, int col, T data) {
+    template <typename T> void set(size_t row, size_t col, T data) {
         assert(sizeof(T) == m_bitdepth / 8);
         if (row < 0 || row >= m_rows || col < 0 || col >= m_cols) {
             throw std::out_of_range("Invalid row or column index");
@@ -33,10 +33,10 @@ class Frame {
         std::memcpy(m_data + (row * m_cols + col) * (m_bitdepth / 8), &data, m_bitdepth / 8);
     }
 
-    ssize_t rows() const { return m_rows; }
-    ssize_t cols() const { return m_cols; }
-    ssize_t bitdepth() const { return m_bitdepth; }
-    ssize_t size() const { return m_rows * m_cols * m_bitdepth / 8; }
+    size_t rows() const { return m_rows; }
+    size_t cols() const { return m_cols; }
+    size_t bitdepth() const { return m_bitdepth; }
+    size_t size() const { return m_rows * m_cols * m_bitdepth / 8; }
     std::byte *data() const { return m_data; }
 
     Frame &operator=(const Frame &other) {
@@ -80,7 +80,7 @@ class Frame {
     }
 
     template <typename T> NDView<T> view() {
-        std::vector<ssize_t> shape = {m_rows, m_cols};
+        std::vector<ssize_t> shape = {static_cast<ssize_t>(m_rows), static_cast<ssize_t>(m_cols)};
         T *data = reinterpret_cast<T *>(m_data);
         return NDView<T>(data, shape);
     }
