+++ core/src/Frame.cpp
@@ -13,7 +13,7 @@ namespace aare {
  * @param cols number of columns
  * @param bitdepth bitdepth of the pixels
  */
-Frame::Frame(std::byte *bytes, ssize_t rows, ssize_t cols, ssize_t bitdepth)
+Frame::Frame(std::byte *bytes, size_t rows, size_t cols, size_t bitdepth)
     : m_rows(rows), m_cols(cols), m_bitdepth(bitdepth), m_data(new std::byte[rows * cols * bitdepth / 8]) {
 
     std::memcpy(m_data, bytes, rows * cols * bitdepth / 8);
@@ -26,7 +26,7 @@ Frame::Frame(std::byte *bytes, ssize_t rows, ssize_t cols, ssize_t bitdepth)
  * @param bitdepth bitdepth of the pixels
  * @note the data is initialized to zero
  */
-Frame::Frame(ssize_t rows, ssize_t cols, ssize_t bitdepth)
+Frame::Frame(size_t rows, size_t cols, size_t bitdepth)
     : m_rows(rows), m_cols(cols), m_bitdepth(bitdepth), m_data(new std::byte[rows * cols * bitdepth / 8]) {
 
     std::memset(m_data, 0, rows * cols * bitdepth / 8);
@@ -39,7 +39,7 @@ Frame::Frame(ssize_t rows, ssize_t cols, ssize_t bitdepth)
  * @return pointer to the pixel
  * @note the user should cast the pointer to the appropriate type
  */
-std::byte *Frame::get(int row, int col) {
+std::byte *Frame::get(size_t row, size_t col) {
     if (row < 0 || row >= m_rows || col < 0 || col >= m_cols) {
         std::cerr << "Invalid row or column index" << '\n';
         return nullptr;
